// src/shared/ui/Text/types.ts
export type TextTheme = 'display-l'
  | 'display-m'
  | 'display-s'
  | 'headline-l'
  | 'headline-m'
  | 'headline-s'
  | 'title-l'
  | 'title-m'
  | 'title-s'
  | 'label-l'
  | 'label-m'
  | 'label-s'
  | 'body-l'
  | 'body-m'
  | 'body-s'

export type TextColor = 'on-primary'
  | 'primary'
  | 'on-secondary'
  | 'secondary'
  | 'on-tertiary'
  | 'tertiary'
  | 'on-surface'
  | 'on-surface-variant'
  | 'error'
  | 'on-error'

export type TextAlign = 'center' | 'inherit' | 'justify' | 'left' | 'right'

export type TextSize = 'display-l'
  | 'display-m'
  | 'display-s'
  | 'headline-l'
  | 'headline-m'
  | 'headline-s'
  | 'title-l'
  | 'title-m'
  | 'title-s'
  | 'label-l'
  | 'label-m'
  | 'label-s'
  | 'body-l'
  | 'body-m'
  | 'body-s'

export type TextType = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'p' | 'span' | 'label' | 'div'

export type TextLineHeight = 'lh-110' | 'lh-120' | 'lh-125' | 'lh-130'

export type TextFontWeight = '400' | '500' | '600' | '700'